@@ -1,6 +1,6 @@
 <!-- Bear.svelte -->
 <script lang="ts">
-	export let maskID: string;
+	let { maskID }: { maskID: string } = $props();
 	export const optionValue = 'Bear';
 
 	function generateUID() {
