@@ -1,36 +1,36 @@
 <!-- Graphics.svelte -->
 <script lang="ts">
-	import { GraphicOption } from '$lib/components/avataaar/options/options.js';
-	import Skull from '$lib/components/avataaar/avatar/clothes/Graphics/Skull.svelte';
-	import Bat from '$lib/components/avataaar/avatar/clothes/Graphics/Bat.svelte';
-	import Cumbia from '$lib/components/avataaar/avatar/clothes/Graphics/Cumbia.svelte';
-	import Deer from '$lib/components/avataaar/avatar/clothes/Graphics/Deer.svelte';
-	import Diamond from '$lib/components/avataaar/avatar/clothes/Graphics/Diamond.svelte';
-	import Hola from '$lib/components/avataaar/avatar/clothes/Graphics/Hola.svelte';
-	import Pizza from '$lib/components/avataaar/avatar/clothes/Graphics/Pizza.svelte';
-	import Resist from '$lib/components/avataaar/avatar/clothes/Graphics/Resist.svelte';
-	import Selena from '$lib/components/avataaar/avatar/clothes/Graphics/Selena.svelte';
-	import Bear from '$lib/components/avataaar/avatar/clothes/Graphics/Bear.svelte';
-	import SkullOutline from '$lib/components/avataaar/avatar/clothes/Graphics/SkullOutline.svelte';
-	import Selector from '$lib/components/avataaar/options/Selector.svelte';
+	import { GraphicOption } from '../../../options/options.js';
+	import Skull from './Skull.svelte';
+	import Bat from './Bat.svelte';
+	import Cumbia from './Cumbia.svelte';
+	import Deer from './Deer.svelte';
+	import Diamond from './Diamond.svelte';
+	import Hola from './Hola.svelte';
+	import Pizza from './Pizza.svelte';
+	import Resist from './Resist.svelte';
+	import Selena from './Selena.svelte';
+	import Bear from './Bear.svelte';
+	import SkullOutline from './SkullOutline.svelte';
+	import Selector from '../../../options/Selector.svelte';
 
-	export let maskID: string;
+	let { maskID }: { maskID: string } = $props();
 
 	const option = GraphicOption;
-	const defaultOption = { component: Skull, optionValue: Skull.optionValue };
+	const defaultOption = { component: Skull, optionValue: 'Skull' };
 
 	const components = [
-		{ component: Bat, optionValue: Bat.optionValue },
-		{ component: Cumbia, optionValue: Cumbia.optionValue },
-		{ component: Deer, optionValue: Deer.optionValue },
-		{ component: Diamond, optionValue: Diamond.optionValue },
-		{ component: Hola, optionValue: Hola.optionValue },
-		{ component: Pizza, optionValue: Pizza.optionValue },
-		{ component: Resist, optionValue: Resist.optionValue },
-		{ component: Selena, optionValue: Selena.optionValue },
-		{ component: Bear, optionValue: Bear.optionValue },
-		{ component: SkullOutline, optionValue: SkullOutline.optionValue },
-		{ component: Skull, optionValue: Skull.optionValue }
+		{ component: Bat, optionValue: 'Bat' },
+		{ component: Cumbia, optionValue: 'Cumbia' },
+		{ component: Deer, optionValue: 'Deer' },
+		{ component: Diamond, optionValue: 'Diamond' },
+		{ component: Hola, optionValue: 'Hola' },
+		{ component: Pizza, optionValue: 'Pizza' },
+		{ component: Resist, optionValue: 'Resist' },
+		{ component: Selena, optionValue: 'Selena' },
+		{ component: Bear, optionValue: 'Bear' },
+		{ component: SkullOutline, optionValue: 'SkullOutline' },
+		{ component: Skull, optionValue: 'Skull' }
 	];
 </script>
 
