@@ -1,7 +1,12 @@
 <!-- Pale.svelte -->
 <script lang="ts">
-	export let maskID: string;
-	export let optionValue: string = 'Pale';
+	let {
+		maskID,
+		optionValue = 'Pale'
+	}: {
+		maskID: string;
+		optionValue?: string;
+	} = $props();
 </script>
 
 <g id={'Skin/' + optionValue} mask={'url(#' + maskID + ')'} fill="#FFDBB4">
