@@ -1,27 +1,46 @@
 <!-- Piece.svelte -->
 <script lang="ts">
-	import { onMount, setContext } from 'svelte';
+	import { onMount, setContext, $effect } from 'svelte';
 	import { allOptions, OptionContext } from './options/options.js';
 
 	// Props
-	export let avatarStyle = '';
-	export let style = {};
-	export let pieceType = '';
-	export let pieceSize = '';
-	export let viewBox = '';
-
-	export let topType = '';
-	export let accessoriesType = '';
-	export let hairColor = '';
-	export let facialHairType = '';
-	export let facialHairColor = '';
-	export let clotheType = '';
-	export let clotheColor = '';
-	export let graphicType = '';
-	export let eyeType = '';
-	export let eyebrowType = '';
-	export let mouthType = '';
-	export let skinColor = '';
+	let {
+		avatarStyle = '',
+		style = {},
+		pieceType = '',
+		pieceSize = '',
+		viewBox = '',
+		topType = '',
+		accessoriesType = '',
+		hairColor = '',
+		facialHairType = '',
+		facialHairColor = '',
+		clotheType = '',
+		clotheColor = '',
+		graphicType = '',
+		eyeType = '',
+		eyebrowType = '',
+		mouthType = '',
+		skinColor = ''
+	}: {
+		avatarStyle?: string;
+		style?: any;
+		pieceType?: string;
+		pieceSize?: string;
+		viewBox?: string;
+		topType?: string;
+		accessoriesType?: string;
+		hairColor?: string;
+		facialHairType?: string;
+		facialHairColor?: string;
+		clotheType?: string;
+		clotheColor?: string;
+		graphicType?: string;
+		eyeType?: string;
+		eyebrowType?: string;
+		mouthType?: string;
+		skinColor?: string;
+	} = $props();
 
 	// OptionContext
 	let optionContext = new OptionContext(allOptions);
@@ -30,7 +49,7 @@
 	setContext('optionContext', optionContext);
 
 	// Collect props into an object
-	$: props = {
+	let props = $derived({
 		topType,
 		accessoriesType,
 		hairColor,
@@ -43,10 +62,12 @@
 		eyebrowType,
 		mouthType,
 		skinColor
-	};
+	});
 
 	// Update optionContext when props change
-	$: updateOptionContext();
+	$effect(() => {
+		updateOptionContext();
+	});
 
 	function updateOptionContext() {
 		const data: { [key: string]: any } = {};
