@@ -1,17 +1,10 @@
 <!-- BlazerSweater.svelte -->
 <script lang="ts">
-	import { onMount } from 'svelte';
-
 	export const optionValue = 'BlazerSweater';
 
-	let path1: string;
-	let mask1: string;
-
-	onMount(() => {
-		const uniqueId = (): string => 'id-' + Math.random().toString(36).substring(2, 9);
-		path1 = uniqueId();
-		mask1 = uniqueId();
-	});
+	const uniqueId = (): string => 'id-' + Math.random().toString(36).substring(2, 9);
+	const path1 = uniqueId();
+	const mask1 = uniqueId();
 </script>
 
 <g id="Clothing/Blazer-+-Sweater" transform="translate(0.000000, 170.000000)">
