@@ -3,7 +3,7 @@
 	import Eyebrow from './eyebrow/eyebrow.svelte';
 	import Eyes from './eyes/eyes.svelte';
 	import Mouth from './mouth/mouth.svelte';
-	import Nose from '$lib/components/avataaar/avatar/face/nose/nose.svelte';
+	import Nose from './nose/nose.svelte';
 </script>
 
 <g id="Face" transform="translate(76.000000, 82.000000)" fill="#000000">
