@@ -1,3 +1,5 @@
 <script lang="ts">
-	export { Default } from '$lib/components/avataaar/avatar/face/nose/Default.svelte';
+	import Default from './Default.svelte';
 </script>
+
+<Default />
